+++ Package.swift
@@ -18,8 +18,8 @@ let package = Package(
     products: [ .library(name: "App", type: .dynamic, targets: ["App"]) ],
     dependencies: [
         .package(url: "https://github.com/fair-ground/Fair", from: "0.6.0"), // must be first
-        .package(url: "https://github.com/jectivex/Jack", from: "2.0.0"),
-        .package(url: "https://github.com/jectivex/JackPot", branch: "HEAD"),
+        .package(url: "https://github.com/jectivex/Jack", from: "2.1.0"),
+        .package(url: "https://github.com/jectivex/JackPot", from: "0.1.0"),
         .package(url: "https://github.com/sqlenclave/SQLEnclave", from: "0.0.1"),
         .package(url: "https://github.com/tiqtiq/WeatherTiq", from: "0.0.1"),
         .package(url: "https://github.com/tiqtiq/LocationTiq", from: "0.0.1"),
@@ -40,7 +40,7 @@ let package = Package(
             .product(name: "Lottie", package: "lottie-ios"),
         ], resources: [
             .process("Resources"), // processed resources
-            .copy("../../App.yml"),
+            .copy("App.yml"),
         ]),
         .testTarget(name: "AppTests", dependencies: ["App"]),
     ]
