--- Sources/App/JackScriptView.swift
@@ -1,7 +1,0 @@
-import FairApp
-
-struct ForecastView : View {
-    var body: some View {
-        wip(JackScriptSampleView())
-    }
-}
