+++ Sources/App/Resources/LICENSE.AGPL
@@ -1,0 +1,1 @@
+../../../LICENSE.AGPL
