+++ Sources/App/WelcomeView.swift
@@ -2,49 +2,39 @@ import FairKit
 
 public struct WelcomeView : View {
     static let introItems = [
-        Tile(id: UUID(uuidString: "98213687-6551-4B64-B9F4-E18EA9479708")!,
-             title: NSLocalizedString("Welcome to **Sun Bow**", bundle: .module, comment: "tile title"),
-             subtitle: NSLocalizedString("Your **forever weather buddy**. Rain or Shine, Sun Bow has got you covered.\n(not literally, of course)", bundle: .module, comment: "tile title"),
+        Card(id: UUID(uuidString: "98213687-6551-4B64-B9F4-E18EA9479708")!,
+             title: NSLocalizedString("Welcome to **Sun Bow**", bundle: .module, comment: "card title"),
+             subtitle: NSLocalizedString("Your **forever weather buddy**!", bundle: .module, comment: "card title"),
              foregroundColor: .init(.white),
              backgroundColors: [.init(.accent)],
              animation: .weatherDayBrokenClouds,
              body: NSLocalizedString("""
-            Welcome to **Sun Bow**. We hope you'll like it here!
-            """, bundle: .module, comment: "tile title")),
-        Tile(id: UUID(uuidString: "F5121E79-2902-4068-98D0-5063A7E317C7")!,
-             title: NSLocalizedString("New Features", bundle: .module, comment: "tile title"),
-             subtitle: NSLocalizedString("The weather never sleeps. Neither do we! This release packs over *twenty-eight* bug fixes and performance improvements.", bundle: .module, comment: "tile title"),
+            Welcome to **Sun Bow**. We hope you'll like it here! Rain or Shine, Sun Bow has got you covered.
+            """, bundle: .module, comment: "card title")),
+        Card(id: UUID(uuidString: "F5121E79-2902-4068-98D0-5063A7E317C7")!,
+             title: NSLocalizedString("New Features", bundle: .module, comment: "card title"),
+             subtitle: NSLocalizedString("The weather never sleeps.\n*Neither do we*.", bundle: .module, comment: "card title"),
              foregroundColor: .init(.white),
              backgroundColors: [.init(.blue)],
              animation: .weatherNightScatteredClouds,
              body: NSLocalizedString("""
-            We are constantly making updates and improvements to Sun Bow.
-            """, bundle: .module, comment: "tile title")),
-        Tile(id: UUID(uuidString: "F0E7A835-E20F-44AA-9BF7-2947569179B3")!,
-             title: NSLocalizedString("Global Database Updated", bundle: .module, comment: "tile title"),
-             subtitle: NSLocalizedString("New cities, updated locations. Browse and search worldwide cities for up-to-the-minute weather data and forecasts. Over 25,000 locations added to the places database.", bundle: .module, comment: "tile title"),
+            We are constantly making updates and improvements to Sun Bow. This release packs over *twenty-eight* bug fixes and performance improvements.
+            We are constantly making updates and improvements to Sun Bow. This release packs over *twenty-eight* bug fixes and performance improvements.
+            We are constantly making updates and improvements to Sun Bow. This release packs over *twenty-eight* bug fixes and performance improvements.
+            """, bundle: .module, comment: "card title")),
+        Card(id: UUID(uuidString: "F0E7A835-E20F-44AA-9BF7-2947569179B3")!,
+             title: NSLocalizedString("Global Database Updated", bundle: .module, comment: "card title"),
+             subtitle: NSLocalizedString("New cities, updated locations.", bundle: .module, comment: "card title"),
              foregroundColor: .init(.white),
              backgroundColors: [.init(.brown)],
              animation: .weatherDaySnow,
              body: NSLocalizedString("""
-            Welcome to **Sun Bow**. We hope you'll like it here!
-            """, bundle: .module, comment: "tile title")),
+            Browse and search worldwide cities for up-to-the-minute weather data and forecasts. Over 25,000 locations added to the places database.
+            """, bundle: .module, comment: "card title")),
     ]
 
     public var body: some View {
-        ScrollView {
-            LazyVStack(alignment: .center, pinnedViews: [.sectionHeaders, .sectionFooters]) {
-                ForEach(Self.introItems.enumerated().array(), id: \.element.id) { index, item in
-                    TileView(item: item.trailing(index % 2 == 0))
-                        .padding()
-                        .background(item.background)
-                        .frame(minHeight: 400)
-                        .cornerRadius(24)
-                        .shadow(radius: 1, x: 2, y: 2)
-                        .padding()
-                }
-            }
-        }
+        CardBoard(cards: Self.introItems)
     }
 }
